+++ far/plugin.hpp
@@ -96,7 +96,7 @@ typedef int (WINAPI *FARAPIMESSAGE)(
     FARMESSAGEFLAGS Flags,
     const wchar_t *HelpTopic,
     const wchar_t * const *Items,
-    int ItemsNumber,
+    size_t ItemsNumber,
     int ButtonsNumber
 );
 
@@ -420,7 +420,7 @@ struct FarListItemData
 
 struct FarList
 {
-	int ItemsNumber;
+	size_t ItemsNumber;
 	struct FarListItem *Items;
 };
 
@@ -630,7 +630,7 @@ typedef int (WINAPI *FARAPIMENU)(
     const struct FarKey *BreakKeys,
     int                *BreakCode,
     const struct FarMenuItem *Item,
-    int                 ItemsNumber
+    size_t              ItemsNumber
 );
 
 
@@ -662,7 +662,7 @@ struct PluginPanelItem
 	const wchar_t *Description;
 	const wchar_t *Owner;
 	const wchar_t * const *CustomColumnData;
-	int           CustomColumnNumber;
+	size_t           CustomColumnNumber;
 	DWORD_PTR     UserData;
 	DWORD         CRC32;
 	DWORD_PTR     Reserved[2];
@@ -2240,11 +2240,11 @@ struct OpenPanelInfo
 	const wchar_t               *Format;
 	const wchar_t               *PanelTitle;
 	const struct InfoPanelLine  *InfoLines;
-	int                          InfoLinesNumber;
+	size_t                       InfoLinesNumber;
 	const wchar_t * const       *DescrFiles;
-	int                          DescrFilesNumber;
+	size_t                       DescrFilesNumber;
 	const struct PanelMode      *PanelModesArray;
-	int                          PanelModesNumber;
+	size_t                       PanelModesNumber;
 	int                          StartPanelMode;
 	enum OPENPANELINFO_SORTMODES StartSortMode;
 	int                          StartSortOrder;
@@ -2329,7 +2329,7 @@ struct SetFindListInfo
 	size_t StructSize;
 	HANDLE hPanel;
 	const struct PluginPanelItem *PanelItem;
-	int ItemsNumber;
+	size_t ItemsNumber;
 };
 
 struct PutFilesInfo
@@ -2337,7 +2337,7 @@ struct PutFilesInfo
 	size_t StructSize;
 	HANDLE hPanel;
 	struct PluginPanelItem *PanelItem;
-	int ItemsNumber;
+	size_t ItemsNumber;
 	int Move;
 	const wchar_t *SrcPath;
 	OPERATION_MODES OpMode;
@@ -2348,7 +2348,7 @@ struct ProcessHostFileInfo
 	size_t StructSize;
 	HANDLE hPanel;
 	struct PluginPanelItem *PanelItem;
-	int ItemsNumber;
+	size_t ItemsNumber;
 	OPERATION_MODES OpMode;
 };
